wish,rekomen,bag,alamat

+++ app/views/daftar/detail.php
@@ -30,18 +30,26 @@
                     </div>
                     <div class="brandescr">
                     <!-- menentukan apakah barang telah di masukkan ke keranjang atau belum -->
+                <?php if(isset($_SESSION['id'])) {?>
                     <?php if($data['cekbag']!=1) {?>
                         <button class="linkks" type="submit" name="barang" value="<?= $data['barang']['Id_Barang'];?>">Masukkan ke Tas</button>
                     <?php }else{?>
                         <div class="linkknot" >Masukkan ke Tas</div>
                     <?php }?>
+                <?php }else{?>
+                    <a class="linkks" href="<?= BASEURL;?>/Regist/pelma" >Masukkan ke Tas</a>
+                <?php }?>
 
                 <!-- menentukan apakah barang telah di masukkan ke wishlist atau belum -->
+                <?php if(isset($_SESSION['id'])) {?>
                     <?php if($data['cekwish']!=1) {?>
                         <a class="logshield" href="<?= BASEURL;?>/Wish/tambWish/<?= $data['barang']['Id_Barang'];?>"><img class="loggo" src="<?= BASEURL;?>/img/wishlist.png" alt=""></a>
                     <?php }else{?>
                         <a class="logshield" href="<?= BASEURL;?>/Wish/hapusdetail/<?= $data['barang']['Id_Barang'];?>"><img class="loggo" src="<?= BASEURL;?>/img/love.png" alt=""></a>
                     <?php }?>
+                <?php }else{?>
+                    <a class="logshield" href="<?= BASEURL;?>/Regist/pelma"><img class="loggo" src="<?= BASEURL;?>/img/wishlist.png" alt=""></a>
+                <?php }?>
                     </div>
                 </form>
 
@@ -55,24 +63,23 @@
             </div>
         </div>
     </div>
-    <div class="main">
-        <div class="barang">
-            <h3 class="bold fontsize">Rekomendasi untuk Anda</h3>
-            <?php foreach($data['barang'] as $it):?>
-            <a href="<?= BASEURL;?>/Main/Detail/<?= $it['Id_Barang'] ?>" class="cardd">
-                <img class="carddis" src="<?= BASEURL."/gambar/".$it['Pic']?>" alt="">
-                <div>
-                    <div class="bris">
-                        <h6 class="brand"><?= $it['Brand']?></h6>
-                        <img  src="" alt="">
-                    </div>
-                    <h6 class="deskripsi"><?= $it['Descr']?></h6>
-                    <h6 class="harga">Rp <?= $it['Harga']?></h6>
+    <div class="bold fontsize">Temuan Teratas & Penawaran Menarik dalam Kategori Wanita</div><br><br>
+    <div class="main1 gray">
+        <?php foreach($data['rek'] as $ss):?>
+                <div class="cardlist">
+                    <a class="cardd" href="<?= BASEURL;?>/Main/Detail/<?= $ss['Id_Barang'] ?>" >
+                        <img class="cardimage" src="<?= BASEURL."/gambar/".$ss['Pic']?>" alt="">
+                        <div class="hades">
+                            <div class="bris">
+                                <span class="brand"><?= $ss['Brand']?></span>
+                                <!-- <button><img class="loggo" src="<?= BASEURL;?>/img/wishlist.png" alt=""></button> -->
+                            </div>
+                            <h6 class="deskripsi"><?= $ss['Descr']?></h6>
+                            <h6 class="harga">Rp <?= number_format($ss['Harga'],0,",",".");?></h6>
+                        </div>
+                    </a>
                 </div>
-                <!-- <a href="">Masukkan ke Tas</a> -->
-            </a>
             <?php endforeach;?>
-            
-        </div>
-    </div>
+    </div><br><br>
+
 </div>
